+++ src/app/sitemap.ts
@@ -116,6 +116,7 @@ const pages = [
   { path: '/resources/public-safety-software-liberia', priority: 0.75 },
   { path: '/resources/public-safety-software-guinea-bissau', priority: 0.75 },
   { path: '/resources/public-safety-software-madagascar', priority: 0.75 },
+  { path: '/resources/public-safety-software-burundi', priority: 0.75 },
     { path: '/resources/public-safety-software-central-african-republic', priority: 0.75 },
     { path: '/resources/public-safety-software-mauritania', priority: 0.75 },
     { path: '/resources/public-safety-software-eritrea', priority: 0.75 },